Remove `rel_max` feature, move `compute_norms`
- Doctests are now working
